@@ -21,22 +21,22 @@ import {MatAutocompleteModule} from '@angular/material/autocomplete';
 import { GraphQLDetailComponent } from './pages/GraphQL/detail/graph-ql-detail';
 import { GraphQLComponent } from './pages/GraphQL/graph-ql-component';
 import { GraphQLLinkComponent } from './pages/GraphQL/links/links';
-import { GraphQLProperties } from './pages/GraphQL/properties/properties';
+import { Properties } from './pages/GraphQL/properties/properties';
 import { GraphQLTable } from './pages/GraphQL/table/table';
 //import {EditDialog} from './pages/applications/applications'
 import {SelectDialog} from './pages/GraphQL/select-dialog/select-dialog';
 import { Applications } from './pages/applications/applications';
 import { LogsComponent } from './pages/logs/logs';
-import { UserEntryDialog } from './pages/user-management/users/dialog/user-dialog';
-import { UserComponent } from './pages/user-management/users/users';
-import { SeverityPickerModule } from './shared/severity-picker/severity-picker';
+//import { UserEntryDialog } from './pages/user-management/users/dialog/user-dialog';
+//import { UserComponent } from './pages/user-management/users/users';
+import { SeverityPicker } from './shared/severity-picker/severity-picker';
 
 @NgModule({
   	declarations: [
-		/*EditDialog,*/ SelectDialog, Applications, LogsComponent,
+		/*EditDialog, SelectDialog, Applications, LogsComponent,
 		GraphQLTable, GraphQLComponent, GraphQLDetailComponent, GraphQLLinkComponent, GraphQLProperties,
-		UserComponent, UserEntryDialog ],
-  	imports: [CommonModule, SeverityPickerModule, MatAutocompleteModule,
+		UserComponent, UserEntryDialog*/ ],
+  	imports: [CommonModule, SeverityPicker, MatAutocompleteModule,
 	  FormsModule, ReactiveFormsModule,
 	  MatButtonModule, MatCheckboxModule, MatCardModule,MatChipsModule, MatDatepickerModule, MatDialogModule, MatDividerModule, MatFormFieldModule, MatInputModule, MatIconModule, MatNativeDateModule, MatSelectModule, MatSortModule, MatTabsModule, MatTableModule, MatToolbarModule
 	],
