@@ -1,5 +1,6 @@
 import { filter } from 'rxjs/operators';
 import { Component, AfterViewInit, OnInit, OnDestroy, Inject, ViewChild, Input, AfterContentInit, ContentChildren, QueryList, ContentChild } from '@angular/core';
+import { CommonModule } from '@angular/common';
 import {ActivatedRoute, NavigationEnd, Params, Router, RouterModule, Routes} from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
 import { Sort } from '@angular/material/sort';
@@ -14,23 +15,23 @@ import {StringUtils} from '../../../utilities/StringUtils'
 import { SelectDialog } from '../select-dialog/select-dialog';
 import { MetaObject } from '../../../utilities/JsonUtils';
 
-@Component( { selector: 'graph-ql-links', templateUrl: 'links.html'} )
-export class GraphQLLinkComponent implements OnDestroy, OnInit, AfterViewInit
-{
+@Component( {
+    selector: 'graph-ql-links',
+    templateUrl: 'links.html',
+    imports: [CommonModule]
+})
+export class GraphQLLinkComponent implements OnDestroy, OnInit, AfterViewInit{
 	constructor( private route: ActivatedRoute, private router:Router, private dialog : MatDialog, private componentPageTitle:ComponentPageTitle, @Inject('IGraphQL') private graphQL: IGraphQL, @Inject('IProfile') private profileService: IProfile, @Inject('IErrorService') private cnsle: IErrorService )
 	{}
 
 	ngOnDestroy(){ this.profile.save(); }
-	ngOnInit()
-	{// @ts-ignore
+	ngOnInit(){// @ts-ignore
 		this.router.events.pipe( filter(e=>e instanceof NavigationEnd) ).subscribe( this.onNavigationEnd );
 	}
-	async ngAfterViewInit()
-	{
+	async ngAfterViewInit(){
 		let self = this;
 		await this.profile.loadedPromise;
-		try
-		{
+		try{
 			let mutations = await this.graphQL.mutations();
 			const mutationA = `add${this.parentType}${this.schema.typeName}`;
 			const mutationB = `add${this.schema.typeName}${this.parentType}`;
@@ -44,35 +45,27 @@ export class GraphQLLinkComponent implements OnDestroy, OnInit, AfterViewInit
 			// for( var f of this.schema.fields )
 			// 	console.log( f );
 		}
-		catch( e )
-		{
+		catch( e ){
 			this.cnsle.error( "could not load page.", e );
 		}
 		this.viewPromise = Promise.resolve( true );
 	}
-	onNavigationEnd =( val:NavigationEnd )=>
-	{
-	}
-	edit( fieldName:string, element )
-	{
+	onNavigationEnd =( val:NavigationEnd )=>{}
+	edit( fieldName:string, element ){
 		var field = this.schema.fields.find( (x)=>x.name==fieldName );
-		this.graphQL.schema( [field.type.underlyingName] ).then( (x)=>
-		{
-			const dialogRef = this.dialog.open( SelectDialog,
-			{
+		this.graphQL.schema( [field.type.underlyingName] ).then( (x)=>{
+			const dialogRef = this.dialog.open( SelectDialog,{
 				width: '600px',
 				height: '650px',
 				data: { schema: x[0], selectedIds:element[fieldName], query:fieldName, mutation:this.mutation, linkTo:this.parent.id, linkToField: this.parentTypeField, subTo:element.id ?? `"${element[ element.target ? "target" : "name" ]}"`, subToField: element.id ? this.schema.subType.idReferenceName : element.target ? "target" : "name", title:StringUtils.capitalize(fieldName), includeDeleted: true, graphQL:this.graphQL }
 			});
-			dialogRef.afterClosed().subscribe( result =>
-			{
+			dialogRef.afterClosed().subscribe( result =>{
 				if( result )
 					this.load();
 			});
 		}).catch( (e)=>console.error(e) );
 	}
-	async load()
-	{
+	async load(){
 		this.viewPromise = null;
 		const valueMember = this.schema.objectCollectionName;
 		let ql = `${this.parentSelect}(filter:{ id:{eq:${this.parent.id}}}){ ${valueMember}{${this.schema.columns}} }`;
@@ -82,36 +75,29 @@ export class GraphQLLinkComponent implements OnDestroy, OnInit, AfterViewInit
 		this.viewPromise = Promise.resolve( true );
 	}
 	cellClick( row:any ){  this.selection = this.selection == row ? null : row; }
-	addLink()
-	{
-		let show = (schema:Table)=>
-		{
+	addLink(){
+		let show = (schema:Table)=>{
 			let selectedIds = this.items.map( (x)=>x.id );
-			const dialogRef = this.dialog.open( SelectDialog,
-			{
+			const dialogRef = this.dialog.open( SelectDialog,{
 				width: '600px',
 				height: '650px',
 				data: { selectedIds:selectedIds, schema:schema, mutation:this.mutation, linkTo:this.parent.id, linkToField: this.parentTypeField, title:`${this.parent.name} ${this.schema.display}` }
 			});
-			dialogRef.afterClosed().subscribe( result =>
-			{
+			dialogRef.afterClosed().subscribe( result =>{
 				if( result )
 					this.load();
 			});
 		};
-		if( this.schema.subType )
-		{
-			this.graphQL.schema( [this.schema.subType.typeName] ).then( (x)=>
-			{
+		if( this.schema.subType ){
+			this.graphQL.schema( [this.schema.subType.typeName] ).then( (x)=>{
 				show( x[0] );
 			}).catch( (e)=>console.error(e) );
 		}
 		else
 			show( this.schema );
 	}
-	removeLink()
-	{
-		let ql = `{remove${this.mutation}("input":{ "${this.parentTypeField}": ${this.parent.id}, `;
+	removeLink(){
+		let ql = `{remove${this.mutation}(${this.parentTypeField}: ${this.parent.id}, `;
 		if( this.selection.id )
 			ql+=`"${new MetaObject(this.schema.typeName).singular}Id": ${this.selection.id}} ) }`;
 		else if( this.selection.target )
@@ -119,16 +105,14 @@ export class GraphQLLinkComponent implements OnDestroy, OnInit, AfterViewInit
 		else
 			ql+=`"name": "${this.selection.name}" })}`;
 
-		this.graphQL.mutation( ql ).then( (x)=>
-		{
+		this.graphQL.mutation( ql ).then( (x)=>{
 			this.selection = null;
 			this.load();
 		} ).catch( (e)=>{this.cnsle.error(e);} );
 	}
 	sortData(event)
-	{
+	{}
 
-	}
 	get haveSelection():boolean{ return !!this.selection; }
 	selection:any|null|undefined;
 	@Input() schema:Table;
@@ -137,7 +121,7 @@ export class GraphQLLinkComponent implements OnDestroy, OnInit, AfterViewInit
 	@Input() parent:any;
 	get id(){ return  this.router.url.substring( 0, this.router.url.lastIndexOf('/')+1 )+this.schema.objectReferenceName; }
 
-	get profile():Settings<PageSettings>{ return this.#profile ?? ( this.#profile = new Settings<PageSettings>(PageSettings, this.id, this.profileService) ); } #profile:Settings<PageSettings>;
+	get profile():Settings<UserSettings>{ return this.#profile ?? ( this.#profile = new Settings<UserSettings>(UserSettings, this.id, this.profileService) ); } #profile:Settings<UserSettings>;
 	get sort(){ return this.profile.value.sort; }
 	get displayedColumns():Field[]{ return this.schema.fields.filter( (x)=>x.displayed && !["created", "updated", "deleted", "target"].includes(x.name) ); }
 	get displayedColumnNames(){ return this.displayedColumns.map( (x)=>x.name ); };// = ["name","target","description","authenticator", "deleted"];
@@ -154,8 +138,7 @@ export class GraphQLLinkComponent implements OnDestroy, OnInit, AfterViewInit
 	showDeleted=false;
 }
 
-class PageSettings
-{
-	assign( value:PageSettings ){   }
+class UserSettings{
+	assign( value:UserSettings ){   }
 	sort:Sort = {active: "name", direction: "asc"};
 }
